# swagger.yaml
swagger: '2.0'
info:
  description: TutorKrub API eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWQiOiI2MTYwNGNhOGI1M2E2OGUzY2FiMmY3MGQiLCJpYXQiOjE2NDk3MjcwMDIsImV4cCI6MTY1MDMzMTgwMn0.KIEC4HoCVpS_1DyVVresBdRHK5n5YmD44Tj1HetPgv4
  version: 1.0.0
  title: TutorKrub API
  contact:
    email: support@tutorkrub.com
# host: petstore.swagger.io
# basePath: /v2
...
paths:
...
  /tutors:
    get:
      tags:
        - tutor
      security:
        - api_key: [ ]
      parameters: [ ]
      responses:
        200:
          description: success
        401:
          description: unauthorized
    post:
      tags:
        - tutor
      security:
        - api_key: [ ]
      parameters:
        - in: body
          name: body
          schema:
            $ref: '#/definitions/TutorRegistration'
      responses:
        200:
          description: ok
    put:
      tags:
        - tutor
      security:
        - api_key: [ ]
      parameters:
        - in: body
          name: body
          schema:
            $ref: '#/definitions/Tutor'
      responses:
        200:
          description: ok
...
  /courses:
    get:
      tags:
        - course
      security:
        - api_key: [ ]
      parameters: [ ]
      responses:
        200:
          description: success
        401:
          description: unauthorized
    post:
      tags:
        - course
      security:
        - api_key: [ ]
      parameters:
        - in: body
          name: body
          schema:
            $ref: '#/definitions/Course'
      responses:
        200:
          description: ok
        405:
          description: invalid input
...
  /search/tutors:
    get:
      tags:
        - search
      parameters:
        - name: id
          in: query
          type: string
        - name: subject
          in: query
          type: string
        - name: level
          in: query
          type: string
      responses:
        200:
          description: success
          schema:
            type: array
            items:
              $ref: '#/definitions/Search'
  /search/courses:
    get:
      tags:
        - search
      parameters:
        - name: id
          in: query
          type: string
        - name: tutorid
          in: query
          type: string
      responses:
        200:
          description: success
          schema:
            type: array
            items:
              $ref: '#/definitions/Search'
...
securityDefinitions:
  api_key:
    type: apiKey
    name: accesstoken
    in: header
definitions:
...
  Course:
    type: object
    properties:
      id:
        type: string
        format: uuid
      title:
        type: string
      price:
        type: number
...
  TutorRegistration:
    type: object
    properties:
      teachSubjects:
        type: array
        items:
          type: string
      teachLevels:
        type: array
        items:
          type: string
      price:
        type: number
  Tutor:
    type: object
    properties:
      teachSubjects:
        type: array
        items:
          type: string
      teachLevels:
        type: array
        items:
          type: string
      price:
        type: number
      educations:
        type: array
        items:
          $ref: '#/definitions/Education'
      commucationChannels:
        type: array
        items:
          $ref: '#/definitions/CommucationChannel'
      onsiteDistrictCode:
        type: number
      additionalDetails:
        type: array
        items:
          type: string
  Education:
    type: object
    properties:
      level:
        type: string
      school:
        type: string
      gpa:
        type: number
  CommucationChannel:
    type: object
    properties:
      channel:
        type: string
      id:
        type: string
  Search:
    type: object
    properties:
      name:
        type: string
      avatarUrl:
        type: string
      teachSubjects:
        type: array
        items:
          type: string
      teachLevels:
        type: array
        items:
          type: string
      price:
        type: number
      educations:
        type: array
        items:
          $ref: '#/definitions/Education'
      commucationChannels:
        type: array
        items:
          $ref: '#/definitions/CommucationChannel'
      onsiteDistrictCode:
        type: number
      additionalDetails:
        type: array
        items:
          type: string
...
basePath: /api
schemes:
  - http